pyproject.toml
[project]
name = "agent-skills-mcp-server"
version = "0.1.0"
description = "MCP server for managing and executing Agent Skills with governance"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "agent-skills-mcp-server" }
]
readme = "README.md"

dependencies = [
    "fastmcp>=2.14.3,<3.0.0",
    "litellm>=1.81.1,<2.0.0",
    "pydantic>=2.12.5,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "pyyaml>=6.0.3,<7.0.0",
    "boto3>=1.42.32,<2.0.0",  # AWS Bedrock用
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "ruff>=0.9.0,<1.0.0",
]

[project.scripts]
agent-skills-mcp = "agent_skills_mcp.server:main"
check = "scripts.check:main"
fix = "scripts.fix:main"
test = "scripts.test:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_skills_mcp"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function calls in argument defaults
    "C901",   # Too complex
    "UP035",  # Import from collections.abc instead of typing
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
